Fix a historical bug regarding the input model
We found that the iCEL1314 model file in the Github was based on an outdated model table (now named OUTDATED_VERSION_STOP_USING_originalModel.xlsx). This model differs from the default iCEL1314 model on the constraints in two reactions,  RM00112 and RMC0005. Both reactions should be blocked to [0,0] by default but the older version allows free flux. To make sure consistency with the model downloaded from WormFlux website, we updated the input model.